+++ src/main/java/org/example/carservice/dto/response/OrderResponseDto.java
@@ -1,5 +1,6 @@
 package org.example.carservice.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.Data;
@@ -15,5 +16,5 @@ public class OrderResponseDto {
     private List<Long> serviceIds;
     private List<Long> productIds;
     private StatusOrder statusOrder;
-    private double costTotal;
+    private BigDecimal costTotal;
 }
