+++ src/main/java/org/example/carservice/service/OrderService.java
@@ -1,5 +1,6 @@
 package org.example.carservice.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import org.example.carservice.model.Order;
 import org.example.carservice.model.Product;
@@ -13,7 +14,7 @@ public interface OrderService {
 
     Order updateStatus(Order order);
 
-    double getTotalCost(Order order);
+    BigDecimal getTotalCost(Order order);
 
     Order addProduct(Order order, Product product);
